+++ lib/mobility.rb
@@ -102,9 +102,6 @@ module Mobility
     # (see Mobility::Configuration#accessor_method)
     # @!method accessor_method
     #
-    # (see Mobility::Configuration#query_method)
-    # @!method query_method
-
     # (see Mobility::Configuration#new_fallbacks)
     # @!method new_fallbacks
 
@@ -116,7 +113,7 @@ module Mobility
     #
     # (see Mobility::Configuration#default_accessor_locales)
     # @!method default_accessor_locales
-    %w[accessor_method query_method default_backend plugins default_accessor_locales].each do |method_name|
+    %w[accessor_method default_backend plugins default_accessor_locales].each do |method_name|
       define_method method_name do
         config.public_send(method_name)
       end