+++ lib/mobility/plugins/query.rb
@@ -3,14 +3,18 @@ module Mobility
   module Plugins
 =begin
 
-@see {Mobility::Plugins::ActiveRecord::Query}
+@see {Mobility::Plugins::ActiveRecord::Query} or {Mobility::Plugins::Sequel::Query}.
 
 =end
     module Query
       extend Plugin
 
-      default true
+      default :i18n
       depends_on :backend, include: :before
+
+      def query_method
+        (options[:query] == true) ? self.class.defaults[:query] : options[:query]
+      end
     end
 
     register_plugin(:query, Query)